+++ src/torchmetrics/__init__.py
@@ -69,6 +69,7 @@ from torchmetrics.regression import (  # noqa: E402
     MeanAbsolutePercentageError,
     MeanSquaredError,
     MeanSquaredLogError,
+    MinkowskiDistance,
     PearsonCorrCoef,
     R2Score,
     SpearmanCorrCoef,