+++ src/torchmetrics/classification/confusion_matrix.py
@@ -408,3 +408,4 @@ class ConfusionMatrix:
         if task == ClassificationTask.MULTILABEL:
             assert isinstance(num_labels, int)
             return MultilabelConfusionMatrix(num_labels, threshold, **kwargs)
+        return None