+++ src/torchmetrics/classification/recall_at_fixed_precision.py
@@ -35,9 +35,10 @@ from torchmetrics.utilities.enums import ClassificationTask
 
 
 class BinaryRecallAtFixedPrecision(BinaryPrecisionRecallCurve):
-    r"""Compute the highest possible recall value given the minimum precision thresholds provided. This is done by
-    first calculating the precision-recall curve for different thresholds and the find the recall for a given
-    precision level.
+    r"""Compute the highest possible recall value given the minimum precision thresholds provided.
+
+    This is done by first calculating the precision-recall curve for different thresholds and the find the recall for
+    a given precision level.
 
     As input to ``forward`` and ``update`` the metric accepts the following input:
 
@@ -117,9 +118,10 @@ class BinaryRecallAtFixedPrecision(BinaryPrecisionRecallCurve):
 
 
 class MulticlassRecallAtFixedPrecision(MulticlassPrecisionRecallCurve):
-    r"""Compute the highest possible recall value given the minimum precision thresholds provided. This is done by
-    first calculating the precision-recall curve for different thresholds and the find the recall for a given
-    precision level.
+    r"""Compute the highest possible recall value given the minimum precision thresholds provided.
+
+    This is done by first calculating the precision-recall curve for different thresholds and the find the recall for
+    a given precision level.
 
     As input to ``forward`` and ``update`` the metric accepts the following input:
 
@@ -210,9 +212,10 @@ class MulticlassRecallAtFixedPrecision(MulticlassPrecisionRecallCurve):
 
 
 class MultilabelRecallAtFixedPrecision(MultilabelPrecisionRecallCurve):
-    r"""Compute the highest possible recall value given the minimum precision thresholds provided. This is done by
-    first calculating the precision-recall curve for different thresholds and the find the recall for a given
-    precision level.
+    r"""Compute the highest possible recall value given the minimum precision thresholds provided.
+
+    This is done by first calculating the precision-recall curve for different thresholds and the find the recall for
+    a given precision level.
 
     As input to ``forward`` and ``update`` the metric accepts the following input:
 
@@ -306,9 +309,10 @@ class MultilabelRecallAtFixedPrecision(MultilabelPrecisionRecallCurve):
 
 
 class RecallAtFixedPrecision:
-    r"""Compute the highest possible recall value given the minimum precision thresholds provided. This is done by
-    first calculating the precision-recall curve for different thresholds and the find the recall for a given
-    precision level.
+    r"""Compute the highest possible recall value given the minimum precision thresholds provided.
+
+    This is done by first calculating the precision-recall curve for different thresholds and the find the recall for
+    a given precision level.
 
     This function is a simple wrapper to get the task specific versions of this metric, which is done by setting the
     ``task`` argument to either ``'binary'``, ``'multiclass'`` or ``multilabel``. See the documentation of
@@ -341,3 +345,4 @@ class RecallAtFixedPrecision:
             return MultilabelRecallAtFixedPrecision(
                 num_labels, min_precision, thresholds, ignore_index, validate_args, **kwargs
             )
+        return None