+++ src/torchmetrics/functional/text/infolm.py
@@ -70,9 +70,10 @@ class _IMEnum(EnumStr):
 
 
 class _InformationMeasure:
-    """A wrapper class used for the calculation the result of information measure between the discrete reference
-    distributions of predicted and reference sentences. The class also handles input validation for `alpha` and
-    `beta` parameters.
+    """A wrapper class used for the calculation of different information measures.
+
+    This metric can be used to measure the information between the discrete reference distributions of predicted and
+    reference sentences. The class also handles input validation for `alpha` and `beta` parameters.
 
     Args:
         information_measure:
@@ -120,7 +121,9 @@ class _InformationMeasure:
                 f"Parameter `beta` is expected to be float differened from 0 and -1 for {information_measure}."
             )
         if self.information_measure == _IMEnum.AB_DIVERGENCE and (
-            any(not isinstance(p, float) for p in [alpha, beta]) or 0 in [alpha, beta, alpha + beta]
+            alpha is None
+            or beta is None
+            or (any(not isinstance(p, float) for p in [alpha, beta]) or 0 in [alpha, beta, alpha + beta])
         ):
             raise ValueError(
                 "Parameters `alpha`, `beta` and their sum are expected to be differened from 0 for "
@@ -139,8 +142,7 @@ class _InformationMeasure:
 
     @staticmethod
     def _calculate_kl_divergence(preds_distribution: Tensor, target_distribution: Tensor) -> Tensor:
-        """Calculate Kullback-Leibler divergence between discrete distributions of predicted and reference
-        sentences.
+        """Calculate Kullback-Leibler divergence between discrete distributions of predicted and reference sentences.
 
         Args:
             preds_distribution:
@@ -166,10 +168,9 @@ class _InformationMeasure:
             Alpha divergence between discrete distributions of predicted and reference sentences.
         """
         _alpha_denom = self.alpha * (self.alpha - 1)
-        alpha_divergence = (
+        return (
             1 - torch.sum(target_distribution**self.alpha * preds_distribution ** (1 - self.alpha), dim=-1)
         ) / _alpha_denom
-        return alpha_divergence
 
     def _calculate_ab_divergence(self, preds_distribution: Tensor, target_distribution: Tensor) -> Tensor:
         """Calculate AB divergence between discrete distributions of predicted and reference sentences.
@@ -190,8 +191,7 @@ class _InformationMeasure:
         c = torch.log(torch.sum(target_distribution**self.alpha * preds_distribution**self.beta, dim=-1))
         c /= self.alpha * self.beta
 
-        ab_divergence = a + b - c
-        return ab_divergence
+        return a + b - c
 
     def _calculate_beta_divergence(self, preds_distribution: Tensor, target_distribution: Tensor) -> Tensor:
         """Calculate beta divergence between discrete distributions of predicted and reference sentences.
@@ -206,8 +206,7 @@ class _InformationMeasure:
             Beta divergence between discrete distributions of predicted and reference sentences.
         """
         self.alpha = 1.0
-        beta_divergence = self._calculate_ab_divergence(preds_distribution, target_distribution)
-        return beta_divergence
+        return self._calculate_ab_divergence(preds_distribution, target_distribution)
 
     def _calculate_renyi_divergence(self, preds_distribution: Tensor, target_distribution: Tensor) -> Tensor:
         """Calculate Rényi divergence between discrete distributions of predicted and reference sentences.
@@ -221,10 +220,9 @@ class _InformationMeasure:
         Return:
             Rényi divergence between discrete distributions of predicted and reference sentences.
         """
-        renyi_divergence = (
+        return (
             torch.log(torch.sum(target_distribution**self.alpha * preds_distribution ** (1 - self.alpha), dim=-1))
         ) / (self.alpha - 1)
-        return renyi_divergence
 
     @staticmethod
     def _calculate_l1_distance(preds_distribution: Tensor, target_distribution: Tensor) -> Tensor:
@@ -308,8 +306,7 @@ def _get_dataloader(
         An instance of ``torch.utils.data.DataLoader`` used for iterating over examples.
     """
     dataset = TokenizedDataset(input_ids, attention_mask, idf)
-    dataloader = DataLoader(dataset, batch_size=batch_size, num_workers=num_workers)
-    return dataloader
+    return DataLoader(dataset, batch_size=batch_size, num_workers=num_workers)
 
 
 def _get_special_tokens_map(tokenizer: PreTrainedTokenizerBase) -> Dict[str, int]:
@@ -322,18 +319,18 @@ def _get_special_tokens_map(tokenizer: PreTrainedTokenizerBase) -> Dict[str, int
     Return:
         A dictionary containing: mask_token_id, pad_token_id, sep_token_id and cls_token_id.
     """
-    special_tokens_maps = {
+    return {
         "mask_token_id": tokenizer.mask_token_id,
         "pad_token_id": tokenizer.pad_token_id,
         "sep_token_id": tokenizer.sep_token_id,
         "cls_token_id": tokenizer.cls_token_id,
     }
-    return special_tokens_maps
 
 
 def _get_token_mask(input_ids: Tensor, pad_token_id: int, sep_token_id: int, cls_token_id: int) -> Tensor:
-    """Generate a token mask for differentiating all special tokens in the input batch. There are 0s for special
-    tokens and 1s otherwise.
+    """Generate a token mask for differentiating all special tokens in the input batch.
+
+    There are 0s for special tokens and 1s otherwise.
 
     Args:
         input_ids:
@@ -357,8 +354,7 @@ def _get_token_mask(input_ids: Tensor, pad_token_id: int, sep_token_id: int, cls
 def _get_batch_distribution(
     model: PreTrainedModel, batch: Dict[str, Tensor], temperature: float, idf: bool, special_tokens_map: Dict[str, int]
 ) -> Tensor:
-    """Calculate a discrete probability distribution for a batch of examples according to the methodology described
-    in `InfoLM`_.
+    """Calculate a discrete probability distribution for a batch of examples. See `InfoLM`_ for details.
 
     Args:
         model:
@@ -524,8 +520,7 @@ def _infolm_compute(
     preds_distribution = preds_distribution[preds_dataloader.dataset.sorting_indices]
     target_distribution = target_distribution[target_dataloader.dataset.sorting_indices]
     # Calculate information measure
-    infolm_score = information_measure_cls(preds_distribution, target_distribution)
-    return infolm_score
+    return information_measure_cls(preds_distribution, target_distribution)
 
 
 def infolm(