+++ src/torchmetrics/image/sam.py
@@ -28,8 +28,9 @@ if not _MATPLOTLIB_AVAILABLE:
 
 
 class SpectralAngleMapper(Metric):
-    """The metric `Spectral Angle Mapper`_ determines the spectral similarity between image spectra and reference
-    spectra by calculating the angle between the spectra, where small angles between indicate high similarity and
+    """`Spectral Angle Mapper`_ determines the spectral similarity between image spectra and reference spectra.
+
+    It works by calculating the angle between the spectra, where small angles between indicate high similarity and
     high angles indicate low similarity.
 
     As input to ``forward`` and ``update`` the metric accepts the following input