+++ src/torchmetrics/utilities/plot.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 from itertools import product
 from math import ceil, floor, sqrt
-from typing import List, Optional, Sequence, Tuple, Union
+from typing import Any, List, Optional, Sequence, Tuple, Union
 
 import numpy as np
 import torch
@@ -133,10 +133,9 @@ def _get_col_row_split(n: int) -> Tuple[int, int]:
     nsq = sqrt(n)
     if nsq * nsq == n:
         return int(nsq), int(nsq)
-    elif floor(nsq) * ceil(nsq) > n:
+    if floor(nsq) * ceil(nsq) > n:
         return floor(nsq), ceil(nsq)
-    else:
-        return ceil(nsq), ceil(nsq)
+    return ceil(nsq), ceil(nsq)
 
 
 def trim_axs(axs: Union[_AX_TYPE, np.ndarray], nb: int) -> np.ndarray:  # type: ignore[valid-type]
@@ -146,11 +145,11 @@ def trim_axs(axs: Union[_AX_TYPE, np.ndarray], nb: int) -> np.ndarray:  # type: 
     """
     if isinstance(axs, _AX_TYPE):
         return axs
-    else:
-        axs = axs.flat  # type: ignore[union-attr]
-        for ax in axs[nb:]:
-            ax.remove()
-        return axs[:nb]
+
+    axs = axs.flat  # type: ignore[union-attr]
+    for ax in axs[nb:]:
+        ax.remove()
+    return axs[:nb]
 
 
 def plot_confusion_matrix(
@@ -158,10 +157,10 @@ def plot_confusion_matrix(
     add_text: bool = True,
     labels: Optional[List[str]] = None,
 ) -> _PLOT_OUT_TYPE:
-    """Inspired by: https://github.com/scikit-learn/scikit-
-    learn/blob/main/sklearn/metrics/_plot/confusion_matrix.py.
+    """Plot an confusion matrix.
 
-    Plots an confusion matrix
+    Inspired by: https://github.com/scikit-learn/scikit-learn/blob/main/sklearn/metrics/_plot/confusion_matrix.py.
+    Works for both binary, multiclass and multilabel confusion matrices.
 
     Args:
         confmat: the confusion matrix. Either should be an [N,N] matrix in the binary and multiclass cases or an
@@ -213,3 +212,54 @@ def plot_confusion_matrix(
                 ax.text(jj, ii, str(val.item()), ha="center", va="center", fontsize=15)
 
     return fig, axs
+
+
+def plot_binary_roc_curve(
+    tpr: Tensor,
+    fpr: Tensor,
+    ax: Optional[_AX_TYPE] = None,  # type: ignore[valid-type]
+    roc_auc: Optional[Union[float, Tensor]] = None,
+    name: Optional[str] = None,
+    **kwargs: Any,
+) -> _PLOT_OUT_TYPE:
+    """Inspired by: https://github.com/scikit-learn/scikit-learn/blob/main/sklearn/metrics/_plot/roc_curve.py.
+
+    Plots the roc curve
+
+    Args:
+        tpr: Tensor containing the values for True Positive Rate
+        fpr: Tensor containing the values for False Positive Rate
+        ax: Axis from a figure
+        roc_auc: AUROC score (computed separately)
+        name: Custom name to describe the classifier
+        kwargs: additional keyword arguments for line drawing
+
+    Returns:
+        A tuple consisting of the figure and respective ax objects (or array of ax objects) of the generated figure
+
+    Raises:
+        ModuleNotFoundError:
+            If `matplotlib` is not installed
+    """
+    _error_on_missing_matplotlib()
+    fig, ax = plt.subplots() if ax is None else (None, ax)
+
+    if isinstance(roc_auc, Tensor):
+        assert roc_auc.numel() == 1, "roc_auc Tensor must consist of only one element"
+        roc_auc = roc_auc.item()
+
+    line_kwargs = {}
+    if roc_auc is not None and name is not None:
+        line_kwargs["label"] = f"{name} (AUC = {roc_auc:0.2f})"
+    elif roc_auc is not None:
+        line_kwargs["label"] = f"AUC = {roc_auc:0.2f}"
+    elif name is not None:
+        line_kwargs["label"] = name
+
+    line_kwargs.update(**kwargs)
+
+    ax.plot(fpr.detach().cpu(), tpr.detach().cpu(), **line_kwargs)
+    ax.set_xlabel("False Positive Rate")
+    ax.set_ylabel("True Positive Rate")
+
+    return fig, ax