+++ tests/unittests/bases/test_aggregation.py
@@ -33,7 +33,7 @@ class WrappedMinMetric(MinMetric):
     """Wrapped min metric."""
 
     def update(self, values, weights):
-        """only pass values on."""
+        """Only pass values on."""
         super().update(values)
 
 
@@ -41,7 +41,7 @@ class WrappedMaxMetric(MaxMetric):
     """Wrapped max metric."""
 
     def update(self, values, weights):
-        """only pass values on."""
+        """Only pass values on."""
         super().update(values)
 
 
@@ -49,7 +49,7 @@ class WrappedSumMetric(SumMetric):
     """Wrapped min metric."""
 
     def update(self, values, weights):
-        """only pass values on."""
+        """Only pass values on."""
         super().update(values)
 
 
@@ -57,7 +57,7 @@ class WrappedCatMetric(CatMetric):
     """Wrapped cat metric."""
 
     def update(self, values, weights):
-        """only pass values on."""
+        """Only pass values on."""
         super().update(values)
 
 
@@ -83,7 +83,7 @@ class TestAggregation(MetricTester):
 
     @pytest.mark.parametrize("ddp", [False, True])
     def test_aggreagation(self, ddp, metric_class, compare_fn, values, weights):
-        """test modular implementation."""
+        """Test modular implementation."""
         self.run_class_metric_test(
             ddp=ddp,
             metric_class=metric_class,