+++ tests/unittests/nominal/test_tschuprows.py
@@ -43,7 +43,7 @@ _input_logits = Input(
 
 @pytest.fixture()
 def tschuprows_matrix_input():
-    matrix = torch.cat(
+    return torch.cat(
         [
             torch.randint(high=NUM_CLASSES, size=(NUM_BATCHES * BATCH_SIZE, 1), dtype=torch.float),
             torch.randint(high=NUM_CLASSES + 2, size=(NUM_BATCHES * BATCH_SIZE, 1), dtype=torch.float),
@@ -51,7 +51,6 @@ def tschuprows_matrix_input():
         ],
         dim=-1,
     )
-    return matrix
 
 
 def _pd_tschuprows_t(preds, target):
@@ -85,6 +84,8 @@ def _pd_tschuprows_t_matrix(matrix):
     ],
 )
 class TestTschuprowsT(MetricTester):
+    """Test class for `TschuprowsT` metric."""
+
     atol = 1e-5
 
     @pytest.mark.parametrize("ddp", [False, True])